Split the HSM policy gate out of test_slip19.py so the rest runs on the Q
test_slip19.py imported enable_hsm_commands from test_hsm.py for the one
test that needs a policy. That fixture is autouse and skips on the Q, and
importing it re-registers it as autouse here too, so all twelve cases
skipped on the Q simulator -- including the eleven that never touch HSM.

The policy gate now lives in test_slip19_hsm.py with the HSM harness
import; test_slip19.py has no HSM dependency left. On the Q simulator it
passes as is: slp9/slok sit outside the supports_hsm block in usb.py,
ux_show_story maps ENTER to 'y', and the address chunker uses 24-char
groups on the wide screen. Nothing in the firmware changed.
