+++ testing/test_slip19.py
@@ -1,17 +1,14 @@
 # (c) Copyright 2026 by Coinkite Inc. This file is covered by license found in COPYING-CC.
 #
-# SLIP-19 ownership proofs (slp9) and their HSM policy gate.
+# SLIP-19 ownership proofs (slp9) outside HSM mode, and the identifier itself.
 #
-# Run with:  py.test test_slip19.py
+# Nothing here needs HSM, so it runs on the Q as well as the Mk4. The policy gate is in
+# test_slip19_hsm.py, which pulls in the HSM harness and therefore skips on the Q.
 #
-import pytest, struct, json, time
+# Run with:  py.test test_slip19.py --sim
+#
+import pytest, struct, time
 from hashlib import sha256
-from ckcc.protocol import CCProtocolPacker
-
-# The HSM harness fixtures live next door; importing them here makes pytest resolve them.
-# enable_hsm_commands is autouse in test_hsm and must be pulled in explicitly, otherwise these
-# tests only pass when some earlier module happens to have left hsmcmd enabled on the simulator.
-from test_hsm import hsm_reset, hsm_status, start_hsm, enable_hsm_commands
 
 AF_P2WPKH = 0x07
 AF_P2TR = 0x23
@@ -136,33 +133,6 @@ def test_slp9_result_is_only_for_slok(dev, press_select):
 def test_slp9_rejects_junk_path(slp9):
     with pytest.raises(Exception):
         slp9(subpath=b"m/84h/0h/zz/1/0")
-
-
-@pytest.mark.parametrize('policy_paths, subpath, allowed', [
-    (["m/84h/0h/0h/1/*"], SEGWIT_PATH, True),
-    (["m/84h/0h/0h/0/*"], SEGWIT_PATH, False),      # wrong branch
-    (["m/86h/0h/0h/1/*"], TAPROOT_PATH, True),
-    ([], SEGWIT_PATH, False),                        # no slip19_paths => never allowed
-])
-def test_slp9_hsm_path_gate(slp9, start_hsm, hsm_reset, policy_paths, subpath, allowed):
-    # Under a policy, only whitelisted paths may be proven, and the confirmation flag is
-    # permitted because the approved policy is the user's standing consent.
-    policy = dict(warnings_ok=True, rules=[dict(min_pct_self_transfer=99)])
-    if policy_paths:
-        policy['slip19_paths'] = policy_paths
-    start_hsm(policy)
-
-    addr_fmt = AF_P2TR if subpath == TAPROOT_PATH else AF_P2WPKH
-    if allowed:
-        proof = slp9(subpath=subpath, addr_fmt=addr_fmt, flags=FLAG_USER_CONFIRMATION)
-        check_proof_shape(proof, flags=FLAG_USER_CONFIRMATION,
-                          witness_items=1 if subpath == TAPROOT_PATH else 2)
-    else:
-        with pytest.raises(Exception) as ee:
-            slp9(subpath=subpath, addr_fmt=addr_fmt, flags=FLAG_USER_CONFIRMATION)
-        assert 'Not allowed in HSM mode' in str(ee.value)
-
-    hsm_reset()
 
 
 # --- the ownership identifier itself ---------------------------------------------
